server/requirements.txt
Django==5.0.1
djangorestframework==3.15.1
django-cors-headers==4.3.1
pymongo==4.6.1
python-dotenv==1.0.1
